Fix RuntimeWarning: 'artistools.__main__' found in sys.modules after import of package 'artistools', but prior to execution of 'artistools.__main__'; this may result in unpredictable behaviour
