+++ artistools/__init__.py
@@ -25,8 +25,7 @@ from artistools import radfield as radfield
 from artistools import spectra as spectra
 from artistools import transitions as transitions
 from artistools import writecomparisondata as writecomparisondata
-from artistools.__main__ import addargs as addargs
-from artistools.__main__ import main as main
+from artistools.commands import addargs as addargs
 from artistools.configuration import get_config as get_config
 from artistools.configuration import set_config as set_config
 from artistools.estimators import read_estimators as read_estimators
